# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/bag-templates": {
      "post": {
        "operationId": "BagTemplatesController_create",
        "parameters": [
          {
            "name": "X-Client-Transport",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "cookie"
              ]
            },
            "description": "Web panels send `cookie` on /auth/otp/verify, /auth/merchant/login, /auth/admin/login, and /auth/refresh to receive the refresh token ONLY as an httpOnly cookie, stripped out of the JSON body. Omit entirely for the mobile app (reads the refresh token from the body)."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBagTemplateDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BagTemplateDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated, but not permitted (wrong actor type, unapproved merchant, or ownership check failed).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a bag template (the reusable listing template offers are created from).",
        "tags": [
          "offers"
        ]
      },
      "get": {
        "operationId": "BagTemplatesController_list",
        "parameters": [
          {
            "name": "storeId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Client-Transport",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "cookie"
              ]
            },
            "description": "Web panels send `cookie` on /auth/otp/verify, /auth/merchant/login, /auth/admin/login, and /auth/refresh to receive the refresh token ONLY as an httpOnly cookie, stripped out of the JSON body. Omit entirely for the mobile app (reads the refresh token from the body)."
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BagTemplateDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated, but not permitted (wrong actor type, unapproved merchant, or ownership check failed).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List the calling merchant's own bag templates.",
        "tags": [
          "offers"
        ]
      }
    },
    "/api/bag-templates/{id}": {
      "get": {
        "operationId": "BagTemplatesController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Client-Transport",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "cookie"
              ]
            },
            "description": "Web panels send `cookie` on /auth/otp/verify, /auth/merchant/login, /auth/admin/login, and /auth/refresh to receive the refresh token ONLY as an httpOnly cookie, stripped out of the JSON body. Omit entirely for the mobile app (reads the refresh token from the body)."
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BagTemplateDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated, but not permitted (wrong actor type, unapproved merchant, or ownership check failed).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get one of the calling merchant's own bag templates.",
        "tags": [
          "offers"
        ]
      },
      "patch": {
        "operationId": "BagTemplatesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Client-Transport",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "cookie"
              ]
            },
            "description": "Web panels send `cookie` on /auth/otp/verify, /auth/merchant/login, /auth/admin/login, and /auth/refresh to receive the refresh token ONLY as an httpOnly cookie, stripped out of the JSON body. Omit entirely for the mobile app (reads the refresh token from the body)."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBagTemplateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BagTemplateDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated, but not permitted (wrong actor type, unapproved merchant, or ownership check failed).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a bag template.",
        "tags": [
          "offers"
        ]
      },
      "delete": {
        "operationId": "BagTemplatesController_deactivate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Client-Transport",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "cookie"
              ]
            },
            "description": "Web panels send `cookie` on /auth/otp/verify, /auth/merchant/login, /auth/admin/login, and /auth/refresh to receive the refresh token ONLY as an httpOnly cookie, stripped out of the JSON body. Omit entirely for the mobile app (reads the refresh token from the body)."
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BagTemplateDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated, but not permitted (wrong actor type, unapproved merchant, or ownership check failed).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Deactivate a bag template (existing offers unaffected).",
        "tags": [
          "offers"
        ]
      }
    },
...
    "/api/stores": {
      "post": {
        "operationId": "StoresController_create",
        "parameters": [
          {
            "name": "X-Client-Transport",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "cookie"
              ]
            },
            "description": "Web panels send `cookie` on /auth/otp/verify, /auth/merchant/login, /auth/admin/login, and /auth/refresh to receive the refresh token ONLY as an httpOnly cookie, stripped out of the JSON body. Omit entirely for the mobile app (reads the refresh token from the body)."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStoreDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoreDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated, but not permitted (wrong actor type, unapproved merchant, or ownership check failed).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a store for the calling (APPROVED) merchant.",
        "tags": [
          "stores"
        ]
      },
      "get": {
        "operationId": "StoresController_list",
        "parameters": [
          {
            "name": "X-Client-Transport",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "cookie"
              ]
            },
            "description": "Web panels send `cookie` on /auth/otp/verify, /auth/merchant/login, /auth/admin/login, and /auth/refresh to receive the refresh token ONLY as an httpOnly cookie, stripped out of the JSON body. Omit entirely for the mobile app (reads the refresh token from the body)."
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StoreDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated, but not permitted (wrong actor type, unapproved merchant, or ownership check failed).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List the calling merchant's own stores.",
        "tags": [
          "stores"
        ]
      }
    },
    "/api/stores/{id}": {
      "get": {
        "operationId": "StoresController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Client-Transport",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "cookie"
              ]
            },
            "description": "Web panels send `cookie` on /auth/otp/verify, /auth/merchant/login, /auth/admin/login, and /auth/refresh to receive the refresh token ONLY as an httpOnly cookie, stripped out of the JSON body. Omit entirely for the mobile app (reads the refresh token from the body)."
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoreDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated, but not permitted (wrong actor type, unapproved merchant, or ownership check failed).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get one of the calling merchant's own stores.",
        "tags": [
          "stores"
        ]
      },
      "patch": {
        "operationId": "StoresController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Client-Transport",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "cookie"
              ]
            },
            "description": "Web panels send `cookie` on /auth/otp/verify, /auth/merchant/login, /auth/admin/login, and /auth/refresh to receive the refresh token ONLY as an httpOnly cookie, stripped out of the JSON body. Omit entirely for the mobile app (reads the refresh token from the body)."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStoreDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoreDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated, but not permitted (wrong actor type, unapproved merchant, or ownership check failed).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a store the calling (APPROVED) merchant owns.",
        "tags": [
          "stores"
        ]
      }
    },
...
  },
  "info": {
    "title": "Kurtar API",
    "description": "Kurtar — surplus-food marketplace API. The contract for all four\nfrontends (merchant-web, admin-web, the consumer app, and the\nlanding site).\n\n**Auth**: every route requires a bearer access token\n(`Authorization: Bearer <token>`) unless explicitly marked \"no auth\nrequired\" below — obtained via one of the /auth/* endpoints. A web\npanel additionally sends `X-Client-Transport: cookie` on every\nlogin/verify/refresh call to receive the refresh token ONLY as an\nhttpOnly cookie (never in the JSON body) — the mobile app omits\nthis header and reads the refresh token from the response body\ninstead.\n\n**Errors**: every error response is `{statusCode, errorCode,\nmessage}` (ErrorEnvelope, below) — branch on `errorCode`, never on\n`message` (not stable across releases, and may be Turkish or\nEnglish depending on the request).",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ErrorEnvelopeDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400,
            "description": "HTTP status code, duplicated from the response status for convenience."
          },
          "errorCode": {
            "type": "string",
            "example": "VALIDATION_ERROR",
            "description": "A stable, machine-readable identifier for this specific error condition — the field every client should branch on, never `message`."
          },
          "message": {
            "type": "string",
            "example": "overallStars must be between 1 and 5.",
            "description": "Human-readable detail, Turkish or English depending on the request — not stable across releases, do not pattern-match on it."
          }
        },
        "required": [
          "statusCode",
          "errorCode",
          "message"
        ]
      },
...
      "CreateBagTemplateDto": {
        "type": "object",
        "properties": {
          "category": {
            "type": "string",
            "enum": [
              "MEAL",
              "BAKERY",
              "GROCERY",
              "PRODUCE",
              "OTHER"
            ]
          },
          "dietFlags": {
            "type": "array",
            "uniqueItems": true,
            "items": {
              "type": "string",
              "enum": [
                "VEGETARIAN",
                "VEGAN",
                "GLUTEN_FREE",
                "LACTOSE_FREE"
              ]
            }
          },
          "storeId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "allergenDisclaimer": {
            "type": "string"
          },
          "originalValueCentsMin": {
            "type": "number",
            "minimum": 0
          },
          "originalValueCentsMax": {
            "type": "number",
            "minimum": 0
          },
          "priceCents": {
            "type": "number",
            "minimum": 0
          },
          "description": {
            "type": "string"
          }
        },
        "required": [
          "category",
          "storeId",
          "title",
          "allergenDisclaimer",
          "originalValueCentsMin",
          "originalValueCentsMax",
          "priceCents"
        ]
      },
      "BagTemplateDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "storeId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "category": {
            "type": "string",
            "enum": [
              "MEAL",
              "BAKERY",
              "GROCERY",
              "PRODUCE",
              "OTHER"
            ]
          },
          "dietFlags": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "VEGETARIAN",
                "VEGAN",
                "GLUTEN_FREE",
                "LACTOSE_FREE"
              ]
            }
          },
          "allergenDisclaimer": {
            "type": "string"
          },
          "originalValueCentsMin": {
            "type": "number"
          },
          "originalValueCentsMax": {
            "type": "number"
          },
          "priceCents": {
            "type": "number"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "active": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "storeId",
          "title",
          "category",
          "dietFlags",
          "allergenDisclaimer",
          "originalValueCentsMin",
          "originalValueCentsMax",
          "priceCents",
          "active",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateBagTemplateDto": {
        "type": "object",
        "properties": {
          "category": {
            "type": "string",
            "enum": [
              "MEAL",
              "BAKERY",
              "GROCERY",
              "PRODUCE",
              "OTHER"
            ]
          },
          "dietFlags": {
            "type": "array",
            "uniqueItems": true,
            "items": {
              "type": "string",
              "enum": [
                "VEGETARIAN",
                "VEGAN",
                "GLUTEN_FREE",
                "LACTOSE_FREE"
              ]
            }
          },
          "title": {
            "type": "string"
          },
          "allergenDisclaimer": {
            "type": "string"
          },
          "originalValueCentsMin": {
            "type": "number",
            "minimum": 0
          },
          "originalValueCentsMax": {
            "type": "number",
            "minimum": 0
          },
          "priceCents": {
            "type": "number",
            "minimum": 0
          },
          "description": {
            "type": "string"
          },
          "active": {
            "type": "boolean"
          }
        }
      },
...
      "CreateStoreDto": {
        "type": "object",
        "properties": {
          "categoryTags": {
            "type": "array",
            "uniqueItems": true,
            "items": {
              "type": "string",
              "enum": [
                "MEAL",
                "BAKERY",
                "GROCERY",
                "PRODUCE",
                "OTHER"
              ]
            }
          },
          "name": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "district": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "latitude": {
            "type": "number"
          },
          "longitude": {
            "type": "number"
          },
          "coverImageUrl": {
            "type": "string"
          },
          "openingHoursJson": {
            "type": "object"
          }
        },
        "required": [
          "name",
          "address",
          "district",
          "city",
          "latitude",
          "longitude"
        ]
      },
      "StoreDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "merchantId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "district": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "latitude": {
            "type": "number"
          },
          "longitude": {
            "type": "number"
          },
          "coverImageUrl": {
            "type": "string",
            "nullable": true
          },
          "categoryTags": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "MEAL",
                "BAKERY",
                "GROCERY",
                "PRODUCE",
                "OTHER"
              ]
            }
          },
          "openingHoursJson": {
            "type": "object",
            "nullable": true,
            "additionalProperties": true
          },
          "active": {
            "type": "boolean"
          },
          "avgStars": {
            "type": "number"
          },
          "ratingCount": {
            "type": "number"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "merchantId",
          "name",
          "address",
          "district",
          "city",
          "latitude",
          "longitude",
          "categoryTags",
          "active",
          "avgStars",
          "ratingCount",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateStoreDto": {
        "type": "object",
        "properties": {
          "categoryTags": {
            "type": "array",
            "uniqueItems": true,
            "items": {
              "type": "string",
              "enum": [
                "MEAL",
                "BAKERY",
                "GROCERY",
                "PRODUCE",
                "OTHER"
              ]
            }
          },
          "name": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "district": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "latitude": {
            "type": "number"
          },
          "longitude": {
            "type": "number"
          },
          "coverImageUrl": {
            "type": "string"
          },
          "openingHoursJson": {
            "type": "object"
          },
          "active": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}